added AT 04.00 | Folder > Create Folder AT 04.08 | Folder > Create Multiconfiguration project in Folder
